feat(vte): implement remaining §08.8b catalog rows — SGR 53/55/73-75, DECSTR, DECSED/DECSEL, SL/SR, DECRQSS-DECSLRM, XTPUSHSGR/XTPOPSGR
SGR additions:
- Overline (53/55): new Attr variants, CellFlags::OVERLINE (bit 16)
- Super/subscript (73/74/75): mutually exclusive flags (bits 17-18)
- CellFlags widened from u16 to u32 (fills alignment padding, Cell stays 24B)
- WireCellFlags updated to u32 in oriterm_mux protocol

CSI dispatch additions:
- DECSTR (CSI ! p) → handler.reset_state() for soft terminal reset
- DECSED (CSI ? J) / DECSEL (CSI ? K) → same handler as ED/EL
  (DECSCA protection not yet implemented)
- XTPUSHSGR (CSI # {) / XTPOPSGR (CSI # }) → SgrSnapshot stack on
  Grid (max 10 entries), saves/restores flags + fg + bg + extra

DCS addition:
- DECRQSS DECSLRM query (DCS $ q s ST) reports 1-based margins

SL/SR already implemented — added test coverage only.

14 new tests, 15 catalog rows updated to verified.

+++ oriterm_mux/src/protocol/snapshot.rs
@@ -39,9 +39,9 @@ pub enum WireColor {
 
 /// Cell SGR flags as raw bits.
 ///
-/// Maps 1:1 to `oriterm_core::CellFlags` bits. Using raw `u16` avoids
+/// Maps 1:1 to `oriterm_core::CellFlags` bits. Using raw `u32` avoids
 /// coupling the wire format to the bitflags type.
-pub type WireCellFlags = u16;
+pub type WireCellFlags = u32;
 
 /// A single terminal cell on the wire.
 ///
